Nesting of flags through reflection
Multiple units that can take flags can be composed in a command struct.
If multiple units refer the same type (i.e. the client), it will only be
initialized once; all occurances of that type will refer to the same
value.

Interpretation of flags is split into two stages: first all flags are
registered with a single flag.FlagSet, and after that Process() is
called. This call can be used to fill in defaults if needed, initialize
structures, make calls to the VC instance, etc.

See govc/cli/walk.go for more details about this nesting mechanism. The
reason to use reflection instead of simply requiring outer structs to
call functions on inner structs is that reflection allows us to
automatically have multiple instances of the same type be set to the
same pointer value.