Configure parent disk on create
It is not necessary to reconfigure a VM to achieve this. Also, if this
isn't done on create, it is not possible to create VMs using the same
parent disk concurrently, as they need to acquire a write lock on the
disk if it used as primary (and therefore writable) disk.