Add functions to about command
The command walker calls the Register and Process functions on the
command struct. If they are not defined on the command struct itself, it
will call them an embedded struct that implements these functions. In
this case, for the about command, these functions weren't available on
the command struct itself, so calling them effectively calls them on the
ClientFlag struct that it embeds. The pointer to the ClientFlag struct is
still nil at that time in the struct walking process.

Implementing the Register and Process functions on the command struct
itself sidesteps this behavior. This is a suboptimal but unavoidable
side effect of embedding flags.