Make Wait function in property package standalone
This function uses CreateFilter to be selective in the updates it
receives. Filters can only be added to property collectors, not removed.
This means that a property collector is no longer useful after being
used by Wait. This in turn means that it is most convenient to have Wait
create and destroy a property collector for its own use.