Ignore missing environmentBrowser field
The virtual machine managed object returns an error for the
environmentBrowser field when it is marked as template. This change
ignores the environmentBrowser field if mentioned in the set of missing
properties for a VirtualMachine managed object.

Also add a package to stash integration tests in and add an integration
test for this error in particular.

This fixes #242.