Change finder functions to no longer take varargs
This means that these functions can safely return a not found error for
patterns that didn't match any object. It is up to the caller to figure
out what to do in case of such an error. For the govc commands, many
will continue executing for the remaining arguments, but exit with a
non-zero status to indicate failure for one of the arguments.