Only retrieve "currentSession" property
The previous implementation tried to retrieve all properties on the
session manager. The "sessionList" property requires the
"Sessions.TerminateSession" privilege. If the user didn't have this
privilege, this property would be missing from the response and the
response would include a "NoPermission" fault for this property. This
fault was then bubbled up to the user as if the session was invalid.

Change this to only retrieve the "currentSession" property.

Thanks to @skuzye for providing debugging information.

Fixes #251.