Make finder.DatacenterList recurse top-level folders
finder.go's function DatacenterList(...) would only look for a
Datacenter directly under the rootFolder when using the wildcard
path ("*").  When there is instead a user-defined Folder (or Folders) in
which the Datacanter object is contained, we need to look through those
top-level folders to find the Datacenter objects.

Notably, the finder.DefaultDatacenter function was impacted as it would
only search for Datacenters under the top-level folder.

This patch fixes that issue by recursing down through any top-level
folders if present.