Finder: support automatic Folder recursion (#663)
Prior to this change, the Finder behaved like the 'ls' command and
as such required a Folder name or wildcard in the path argument to
find resources within folders.  The original "list" mode of the Finder
is maintained, but will now switch to "find" mode if the path argument
does not contain a "/".  See find/doc.go for details.

The public Finder method signatures remain the same, with the exception of
ManagedObjectListChildren which now takes an optional (variadic) list of
types.

No new public methods or types have been added.

The 'list.Recurser' type is no longer exported and has moved to an unexported
type within the find package.  This part of the change assumes that 'list.Recurser'
was not used directly outside of the govmomi repo.

The tests depend on the Go based vCenter simulator, which currently lives
in the vmware/vic repo, but will be moving to the govmomi repo at some point.

Fixes #190
Fixes #519
Fixes #540

Closes #507
Closes #593