toolbox: SendGuestInfo before the vmx asks us to
We send the IP (and NIC info) in response to the broadcastIP request.
There can be ~30 seconds between the Reset request (1st call) and the broadcastIP request.
For existing VMs, the IP/NIC info is cached in the vmdb and doesn't take as long.

With this change, the IP/NIC info is pushed asap and takes < 1 second

Before:
  guest.toolsStatus  types.VirtualMachineToolsStatus  toolsOk
  Waiting for VM ip from toolbox...30.01

After:
  guest.toolsStatus  types.VirtualMachineToolsStatus  toolsOk
  Waiting for VM ip from toolbox...0.06