toolbox: reset session when invalidated by the vmx
Seen cases where ESX is under heavy load and the toolbox session
cookie is invalidated by the vmx.

This change will backoff the polling loop and also re-open the message
channel to get a new session cookie.

On the guest side, this resulted in a flood of:
   Message: Unable to send a message over the communication channel 0
   Message: Unable to poll for messages over the communication channel 0

On the vmx site (vmware.log), this resulted in a flood of:
   GuestMsg: Channel 0, Protocol error, state: 0
   GuestMsg: channel 0: wrong cookie, discarding message.