Use base type for DVS backing info
The DVS EthernetCardBackingInfo method was originally written to use the
mo.VmwareDistributedVirtualSwitch type due to a limitation that required
the Go type to be the same as the managed object type.

That limitation was fixed in b34f346, allowing for use of base types.

The simulator also uses the base DistributedVirtualSwitch type now to cover that case.

Fixes #843