Add method to find a CustomFieldDef by Key
Add new type 'CustomFieldDefList' of type '[]types.CustomFieldDef'.

Change the 'CustomFieldManager.Field' method signature so it returns
the new 'CustomFieldDefList' type. This makes it possible to attach
methods on top of the 'CustomFieldDefList' to search the list. This is
not a breaking change because the underlying type is still
'[]types.CustomFieldDef'.

Add 'CustomFieldDefList.ByKey' method to find a 'CustomFieldDef' by the
'Key' attribute.

Resolves: #854