import.spec not to assign deploymentOption
We used to forcefully assign deploymentOption among [small, medium, large]
when we call import.spec on an ova/ovf that doesn't include
deploymentOption. This could generate an option file that cannot be
accepted by import.ova or import.ovf. After this change we only assign
deploymentOption when it's included in ova/ovf file.

Fixes #487