Switch to kr/pretty package for the -dump flag
The kr/pretty package is much closer to Go code that can compile compared to go-spew.
I'd been using the modified version of kr/pretty to generate simulator templates in the past.
But, had done with either a throwaway main.go or temporary hacks to existing govc commands.
The *.info and object.collect commands can now be used with the '-dump' flag to generate these templates.