Ignore AcquireLocalTicket errors
If no username is provided, the govc client tries to acquire a local ticket.
When invoked remotely, ESX returns an InvalidRequestFault (VC returns NotSupportedFault).
So, rather than return an error here, fallthrough to Login() with the original User to
to avoid what would be a confusing error message.

Don't try AcquireLocalTicket at all against vCenter as it is unsupported.

Fixes #952