govc: Add vm.vapp.remove
This subcommand removes vApp configuration from a virtual machine.

The current reasoning: Terraform currently lacks vApp configuration
support. This is being worked on in
terraform-providers/terraform-provider-vsphere#303, but in the meantime
I'm currently working on some code (not necessarily related to the work
going on in the provider) that needs to leverage guestinfo in a CoreOS
template. Working with these attributes in extraConfig is not possible
if vApp configuration is present.

This is functional and has been manually tested, but I'm going to look
at adding the rest of the functionality to govc so that vApp properties
can be largely manipulated within govc itself. This will require work to
both govc and the simulator.

Reference: #574