simulator: Add virtual machine vApp configuration support
This adds virtual machine vApp configuration support to the simulator,
supporting all fields that we currently allow with govc (which is pretty
much everything except EULA and unrecognized OVF sections).

Tests still need to be created for this, and tests for the govc
functionality still need to be added to the bats tests.