vcsim: avoid data races
The method call dispatcher locks the receiver object, so methods do not need to explicitly do so.
However, methods that touch other objects need to use the Registry WithLock method or helpers that use it underneath.

We may still need to cover some other paths, such as reading within PropertyCollector.  And we can fine tune to opt-out of
locking where it isn't needed.  But the race detector is happier now as-is.

Fixes #990