Fix dvs.portgroup.info filtering
This regression was introduced by 585cf5e , where the DistributedVirtualSwitch
FetchDVPorts wrapper was added to the object package. The new method did not
include the optional 'Criteria' param, so no filtering was applied to the
dvs.portgroup.info command.  The command would panic on uplink portgroups, which
are normally filtered out by default.  But the same panic would happen 0.15 if
the '-uplinkPort' flag is given.

This change also adds an uplink portgroup to simulator switches, which would
reproduce the issue.

The dvs.portgroup.info command now has support for using the '-json' and '-dump' flags.

Fixes #1059
Fixes #1060