Add -firmware parameter to 'govc vm.create' with values bios|efi
Without this option it is not possible to specify that a VM should use
EFI instead of BIOS as firmware type.

When creating a new VM per default BIOS will be used. This is the same
as before with the difference that now it is explicit (i.e. the
parameter is always passed).

fixes #1071

fix