Refactor LoginExtensionByCertificate tunnel usage
Originally, a call to soap.Client.SetCertificate configured the client's proxy to use the SDK for every request.
However, the SDK tunnel is only required for the LoginExtensionByCertificate call itself.  After we have a session,
the normal path can be used without the SDK tunnel.  Use of the SDK tunnel is now limited to within the SessionManager
LoginExtensionByCertificate wrapper.

- Add govc session.login -extension flag

- Add vcsim LoginExtensionByCertificate implementation

- Add govc extension.setcert -set flag (handy for generating test certs)

- Remove govmomi.Client LoginExtensionByCertificate related methods