Fixed leading "/" requirement in FindByInventoryPath
https://code.vmware.com/apis/358/vsphere#/doc/vim.SearchIndex.html#findByInventoryPath

Finds a managed entity based on its location in the inventory. The path is separated by slashes ('/').
For example, a path should be of the form "My Folder/My Datacenter/vm/Discovered VM/VM1".
A leading slash or trailing slash is ignored.
Thus, the following paths all represents the same object: "a/b", "/a/b", "a/b/", and '/a/b/'.
Slashes in names must be represented using %2f, following the standard URL syntax.
Any object in the inventory can be retrieved using this method, including resource pools and hosts.