simulator: Remaining VM vApp property tests
This provides the coverage for the remaining vApp property support for
virtual machines in the simulator.

In addition to that, EULA support has been added in both the simulator
and govc. I'm thinking about making the CLI support a bit better though
as right now it's currently just a command line argument, which may be
unwieldy with larger EULAs, in addition to the fact that there's no CLI
support right now for *viewing* EULAs.