Allowing the use of STS for exchanging tokens
In the simplest STS is responsible for issuing tokens required for
authentication. In the most common case it issues SAML Bearer token when
client authenticates using credentials and SAML Holder-of-key (HOK) token
when certificate/private key pair is used for authentication.

STS however supports also more sophisticated use-cases for exchanging
one type of token for another. Allowing to exchange an HOK token for
bearer token or use of JWT tokens.

This change adds an option to allow requesting a specific KeyType when
calling Issue to obtain a new token and use a previously obtained Signer
to authenticate the new call to STS to issue another token.