Omit namespace tag in generated method body response types
To workaround a property collector issue where a client needs to invoke a PropertyCollector method with a namespace other than vim25.
In this case, the body Response type namespace will be that of the request (e.g. "urn:vsan"), which Go fails to decode when the Res
field tag has the urn:vim25 namespace.
The response body type is already given as an input to RoundTrip(), so there's no need to validate the namespace on the client side
and decoding works as expected.
Added custom encoding to the simulator to add namespace to the response body type.