vcsim: fork httptest server package
As of Go 1.12, the "httptest.serve" flag is not registered unless it is present in os.Args.
There is no other way to set the httptest.serveFlag variable, breaking vcsim's default setting
of port 8989.

The simulator/internal/{server,testcert}.go files are a copy from Go's 1.11 httptest package.

We can remove "httptest.serve" and replace with vcsim's "l" flag, along with removing other
workarounds for the httptest package in a future PR.

Fixes #1400