govc: fix vm.change against templates
When a VM is marked as a template, only the name and annotation can be changed.
However, the vm.change command was sending an empty "<tools></tools>" field in the request,
causing vCenter to respond with a NotSupported fault.

Updates vcsim to behave as vCenter does in this case.

Fixes #1430