Fix port signature in REST endpoint token auth
When http.Request.URL.Port is not set on the client side, is also omitted from the request "Host" header.
In this case, on the server side, the default port value used in the message signature is "80".
This change fixes the case when URL.Port is explicitly set, which is then included in the "Host" header.