Fix deadlock for keep alive handlers that attempt log in
Over the last couple of months we have seen test time outs in a project
using govmomi at a frequency of a few failures a week. The reason for
those timeouts lies in a deadlock in the govmomi code driving the keep
alive handler.
In particular, after we sent a Login request we start a timer issuing
keep alive callbacks. Similarly, after issuing a Logout request we stop
this timer.
Every so often we attempt a logout and while in the process of doing so
the timer gets a tick and attempts to log in. Because we wait for all
keep alive handlers to return and do so with a lock held that is also
acquired as part of the timer start that is performed as part of the
login, we deadlock.
To fix this issue this change moves the logic to stop the keep alive
timer before the actual logout call, effectively preventing such a
scenario from occurring.