Fix hostsystem ManagementIPs call
Currently hostsystem ManagementIPs chooses 1st CandidateVnic in
hosts Config.VirtualNicManagerInfo.NetConfig property as the mgmt ip.
This is not correct. This changes switches to using  SelectedVnic
property for a management nicType instead.

Testing Done: Added a unittest.
Used govc to upload file to datastore and verified that IP chosen
was based on SelectedVnic property.