simulator: add interface for VirtualDiskManager
This is part of a fix for #1841.

It allows custom implementations of the VirtualDiskManager to be registered into
the registry without disrupting internal use of the corresponding singleton, as
long as they match VirtualDiskManagerInterface.

The methods there are extended to include a *simulator.Context. Even though the
default implementations don't need it, custom implementations might.