simulator: refactor folder children operations
The previous approach had the following issues:
- some private methods were attached to Folder, when they really applied to
  mo.Folder. As a result, they prevented using a custom implementation of a
  simulator folder object
- the putChild/removChild methods didn't lock the folder properly, generating
  data races

To address the former, move from private Folder methods to functions that take
a *mo.Folder as first parameter, and adjust call sites.

To address the latter, we need to add proper locking on the *mo.Folder object.
But unfortunately we need the session locking for these, as these operations
are (also) applied during some simulator.Folder API operations, meaning that the
folder is already locked for these.
As a result, we need to pass a simulator.Context around, and adjust the
corresponding call sites.