Move simulator lookupservice registration into ServiceInstance
Previously, registration was only being done once because the
atomic used was a package variable.  That meant the lookupservice
would only register once for the life of the process.  This isn't
ideal for use in unit-tests which start and stop the simulator
for each test.  Subsequent tests wouldn't register the lookup
service and fail.