Example doc comments update using API docs content
I'm new to the API and this package and am learning my way
around both. The main project README helpfully directs the
reader to the API docs for more information (thanks for that).
That said, is there any interest in adding documentation
directly in this package as doc comments?

This commit (and subsequent PR) is an example of what
I'm referring to.

Thanks in advance for your feedback.

refs https://vdc-download.vmware.com/vmwb-repository/dcr-public/a5f4000f-1ea8-48a9-9221-586adff3c557/7ff50256-2cf2-45ea-aacd-87d231ab1ac7/vim.ManagedEntity.html