Associate every registry lock with a Context.
This allows us to elide locking when the request already holds the lock.

This scenario comes up because any path that takes a registry lock may discover
that the lock was taken by a caller already, which results in deadlock.

Prior to this change, there was a very simple lock elision policy: if the lock
refers to the managed object that is handling the request, it is elided. This
made sense historically, because 1) the handling MO is unconditionally locked
before its method is called, and 2) there are relatively few cases where lock
elision is actually required, so it was good enough.

This change modifies the lock elision policy: if the lock is already held by
the same simulator.Context (read: same request), it is elided.

To accomplish this, WithLock() now requires a simulator.Context. This requires
passing simulator.Context through to functions that previously didn't need it,
which accounts for many of the files touched in this change.

The other notable change is that internalContext has been removed. Since we now
assume Context to be coupled to the request, re-using a Context for multiple
requests violates that assumption. It appears that most uses of internalContext
were out of convenience, so they were easily replaced with spoofContext(),
which generates a new simulator.Context.

One tricky spot is the simulator's RoundTrip() function, which is used when
simulator code wants to use a client for convenience (the finder being the main
example). For this situation, there is now an internalSession that is
associated with the Context created to service the client call. It would be
ideal to somehow link the Context in RoundTrip() to the Context from which the
client was invoked, and there are ways to accomplish this, but the use case is
presently narrow and the risks are not high (see Model create/load).

That said, using client/RoundTrip() should be avoided to keep locking behavior
consistent.