Simulator Task Delay
This change makes it possible to introduce arbitrary delays for simulator
tasks.

There was already a delay feature for the API call itself. While this
faciliates some testing of client timeouts, it was not possible to test the
scenario where the server is responsive, but the task is taking an unusually
long time.

This change adds a global variable, TaskDelay, that can be configured similarly
to the API delay: a universal delay applied to all tasks, and/or per-method
delays.

Tests were added for both API and Task level delays.