Ensure lock hand-off to simulator.Task goroutine
It is reasonable that a handler func might access some data, prepare a
function closure to be the body of a task, then run that task
asynchronously in a goroutine, and return.

Prior to this commit, the async goroutine would acquire its own lock. If
the goroutine runs fast enough, it will acquire the lock before the
calling routine releases its lock, so any data protected by that lock
would not have changed. But if the goroutine does not acquire the lock
fast enough, it's possible someone else gets it and changes data before
the goroutine can get the lock back.

This could lead to subtle bugs in otherwise-reasonable code.

To address this, this change:
- factors WithLock() into 2 functions:
  - WithLock(), which behaves as before, and
  - AcquireLock(), which gets the lock and returns a function that can
    be used to release the lock.
- uses AcquireLock() in the "calling routine"
- calls the release-lock function after the "async goroutine" executes
  the task.

Because locks are shared whenever they're taken by the same Context,
this has the effect of the calling routine "handing off" the lock to the
async task goroutine, eliminating the race described above.