Isolate the SSO bats test that times out
This patch isolates the govc test for sso.bats -- the test that
often times out when run as part of GitHub actions. This change
is to determine whether or not the issue is specific to sso.bats
or if the timeout occurs at that point during the 200+ test run
for the uber-bats test. This patch also excludes the sso.bats
test from the uber run.

To enable sso.bats, it must be run with SSO_BATS=1. The make
target "govc-test-sso" can be used to run sso.bats.