ESX Agent Manager (EAM) Client and Simulator
This patch introduces the bindings, client, and simulator
for the ESX Agent Manager (EAM). EAM is a long-lived service on vCenter
that acts as an intermediary for provisioning agent virtual machines and
VIB modules on behalf of the user.

The following packages include:

  * ./eam           - the EAM client
  * ./eam/types     - the EAM bindings - types
  * ./eam/methods   - the EAM bindings - methods
  * ./eam/object    - the EAM helper objects, ie. more than just MoRefs
  * ./eam/simulator - the EAM simulator

For more information on how to use EAM, please see the public, EAM SDK
at http://bit.ly/eam-sdk.

This patch also updates the vC Simulator code so when containers
are deployed as VMs, goroutines are created to monitor the containers
and automatically update the runtime properties of the VM to match
the underlying state of the container, such as power state, IP addr,
etc.