Add support for golangci-lint
This patch introduces support for golangci-lint, the de facto
linter for Golang code. This patch replaces the former goimports
and govet targets in the Makefile with the lint target. Please
note the old, check target is still present for backwards
compatability.

The golangci-lint binary is obtained automatically by using the
new hack/tools directory pattern. The directory hack/tooks is a
Go sub-module, github.com/vmware/govmomi-tools, which allows it
to have its own go.mod and go.sum files. This way the module can
be used to maintain a list of Go-based tooling for the project
without impacting the actual dependencies in the root Go module.