Support pre-auth handlers in vC Sim
This patch adds support for pre-auth handlers in vC Sim. This
enables SDKs registered with vC Sim to return their own errors
if a session does not exist.

This patch includes a validation of this behavior by asserting
the EAM simulator should return EamInvalidLogin when a session
is not present instead of vim.NotAuthenticated.