Update GitHub Test Action to use Makefile
This patch updates the GitHub Test Action to use the Makefile
for running the tests. New options are included in the Makefile
to make it easy to parameterize certain values, such as:

  * test count
  * test timeout
  * race history size

The Makefile has also been updated so that certain options that
are parameterized independently, ex. TEST_COUNT, will force an
error if a person tries to specify them in the raw via TEST_OPTS.
For example:

    $ TEST_OPTS=-race make test
    Makefile:126: *** The -race flag is enabled by default & cannot be specified in TEST_OPTS.  Stop.

    $ TEST_OPTS="-count 1" make test
    Makefile:122: *** Use TEST_COUNT to override this option.  Stop.

    $ TEST_OPTS="-timeout 5m" make test
    Makefile:130: *** Use TEST_TIMEOUT to override this option.  Stop.