Fix host lookup in sim/vm/hostInMM
This patch addresses an issue with the way the host is looked
up for a VM in the simulator when CustomizeVM_Task is called.
The previous logic used "ctx.Map", which should work, but for
some reason this is resulting in unexpected behavior.
Switching to the "Map" singleton fixes the issue, but it is
something we will want to reexamine later as there is an
effort to move away from this singleton.