fix: avoid use of vNIC IP in guest.TransferURL if there are multiple
InitiateFileTransfer{To,From}Guest methods return an ESX host's inventory name (HostSystem.Name).
This name was used to add the host to vCenter and cannot be changed
(unless the host is removed from inventory and added back with another name).
The name used when adding to VC may not resolvable by this client's DNS, so we prefer an ESX management IP.
However, if there is more than one management vNIC, we don't know which IP(s) the client has a route to.
Leave the hostname as-is in that case or if the env var has disabled the preference.