fix: avoid vNIC mapping in guest.TransferURL if URL.Host is an IP
No need to map the guest transfer URL host to the ESX vNIC IP if the URL.Host is already an IP address.
In which case we can also avoid property collection of the vNIC config.
Using the original IP also a must when the ESXi host is dual stack, but the client has only pure IPv6.

Changed the host cache value from URL.Hostname() to URL.Host, so IP-literal form is maintained.
e.g. URL.Hostname() is "fe80::1" when URL.Host is "[fe80::1]:80"

Fixes #2553