Re-work TenantManager addition to ServiceContent for older clients
Older (Java/Python/non-Go) clients do not have TenantManaager in vmodl API and
will break if they receive TenanatManager entry in ServiceContent.

This patch reworks the addition of TenantManager moref entry in ServiceContent,
by default the TenantManager morer entry in ServiceContent is kept nil to
avoid breaking clients older than 6.9.1 (that version of vmodl does not
define TenantManager). To return a valid TenantManager ref in ServiceConent
in the response to newer clients (either on RetrieveServiceContent() API call
or the property collector call on ServiceInstance object), we would need
to handle client vmodl version properly, expecting this as a separate change,
this patch keeps the TenantManager unwired and skips the new tests.