fix: CHANGELOG sorting and generation
The CHANGELOG.md is now sorted in semver order, i.e. v0.24.2 will be
sorted before v0.25.0 and not before v0.27.1 as before. This fixes
cherry-picks and is more intuitive for the reader. Also, only
semver-compliant tags are included. This is to work around a bug in
git-chlog and also drop any non-semver compliant tags from the log

Closes: #2638 #2655
Signed-off-by: Michael Gasch <mgasch@vmware.com>