vcsim: Fix distribute VMs across resource pools
This fix aligns the behavior of vcsim with the documentation for the
`pool` and `machine` model flags for cluster configurations.

When no child pools are created, i.e. `pool` is not specified, VMs will
be deployed in the root resource pool of a cluster, i.e. `RP0`. The VMs
will also be prefixed with the corresponding resource pool. In order to
not break standalone host behavior, VMs deployed on a standalone host
will not carry a resource pool identifier in their name. See the linked
issue for deployment examples and naming.

The total number of VMs deployed in a VPX model is the result of the
number of datacenters, clusters and resource pools.

BREAKING:

The name of virtual machines deployed in `vcsim` in a cluster (and
optionally child resource pools) has changed to include the
corresponding resource pool name. VM names deployed to standalone hosts
in `vcsim` are not changed.

Closes: #2047
Signed-off-by: Michael Gasch <mgasch@vmware.com>