fix: avoid possible panic in HostSystem.ManagementIPs
If the HostSystem.Config or HostSystem.Config.VirtualNicManagerInfo fields are nil, the ManagementIPs method would panic.
I was not able to reproduce this situation in a real VC, even with disconnecting a host from VC.
Assuming this state is possible while an ESX host is being upgraded. Updated the test to nil both fields to ensure we don't panic.