vcsim: Remove VM Guest.Net entry when removing Ethernet card
When an Ethernet card is added, an entry for the card is appended to
the VMs Guest.Net. If the Ethernet card is later removed, this entry
was not removed. If Customize() is later called, it will fail with a
NicSettingMismatch error because the Guest.Net still contains an entry
for the removed card.

In configureDevice() always append to the VM's Guest.Net so it is kept
consistent because vcsim Edit operation is done by removing and then
adding the device.