api: Generate JSON tags
This patch updates the type generation code to support adding
JSON tags to generated type field data.

BREAKING:

Updating the `vim25/types` with camelCased, JSON tags affects
the output of the `govc` command when using the `-json` flag
to emit results as JSON. Field names that used to be PascalCased
will now be camelCased. Queries for tools that parse JSON, such
as `jq`, will need to be updated to use the new camelCased fields.