fix: Heal the broken Namespace API
This patch fixes a few bugs related to the Namespaces API. There
was an NPE occurring due to a pointer variable being used prior to
being assigned. The "govc namespace.logs.download" command also
just failed to work due to the VAPI client assuming the response
will always be JSON and not possibly, binary data as one sends when
one receives a tar file. Finally, the namespace.bats file used to
use the cluster name "Workload-Cluster", but due to a hack in the
simulated namespaces API, it only recognizes clusters that begin
with "WCP".