chore: JSON Encoding w Discriminator Support
This patch introduces support for discriminators to the "encoding/json" package.
A discriminator is a field in a JSON object that indicates the object's type.
For example:

        {
            "breed": "Lab"
        }

The above JSON does not indicate the Go type into which the object should be
decoded. However, it is possible to store that Go type information along with
the JSON:

        {
            "_typeName": "Dog",
            "breed": "Lab"
        }

The "_typeName" field teaches the JSON decoder to unmarshal the object into the
Go type named "Dog".

This patch introduces functions for the json.Encoder and json.Decoder that may
be used to active support for discriminators.