chore: Remove JSON tag for "This" / ArrayOf "_value"
This patch removes the JSON tag for properties named "This" of
type "ManagedObjectReference" as VMware's SOAP over JSON project
does not use this.

This patch also ensures any of the "ArrayOf" types use the type
discriminator name "_value" for the name of the type's single
field which is a slice.