chore: Use socat with govc-tests
This patch introduces socat for reading the vC Sim
env var named pipe in the govc-tests. Socat allows
for a timeout if reading the named pipe hangs, which
we suspect may be the cause of the hung tests in the
govc-test GH action workflow. This won't necessarily
fix the tests from failing, but if the named pipe is
the issue, the tests should no longer fail due to a
timeout we have been unable to root cause with any
certainty.