vcsim: support container backing for hosts
This adds support for backing a host with a container in a similar
manner to how we back VMs with containers.

Hosts do not have the VM ExtraConfig mechanism, and "creation" of a
host is more "register the existance of" vs VMs that are actively
constructed from the provided spec. As such, this uses the advanced
options mechanism provided by the per-host Option Manager instead
of ExtraConfig, but following the same "RUN.container" key/value
approach for defining a container backing.

The created container for a host has the following volumes defined:
* bootbank (read-only)
* altbootbank (read-only)
* OS-DATA (read-write)
* datastore1 (read-write)

The volumes have suitably formed UUIDs, are mounted under
/vmfs/volumes, and have symlinked pretty names.

The volumes are associated with the host via labels, allowing the
use of filtered queries to retrieve volumes associated with a given
host.

All docker invocation is kept in container.go and out of the
container_xxx.go files. Not clear this is a fundamental benefit, but
should make it easier if we ever want to support remote docker hosts.