vcsim: per-host OptionManager instances and differentiated roles
Creates an OptionManager instance per-host, with valued seeded from the
ESX template but not directly referencing it, ie. template changes will
not reflect into existing OptionManager instances.
OptionManager Query and Update methods work as expected.

Changes made via OptionManager are reflected into host.Config.Options
array, but it's a unidirectional reflection. This is done to match
infered behaviour of ESX.

There are two OptionManager instances for ESX (and I assume for VC),
For ESX they are found at:
* ServiceContent.setting
* ConfigManager.advancedOptions

The settings for ESX are empty, and the template had named the adv
opts as settings. This adds an empty Setting array in the templates to
clearly differentiate which set of BaseOptionValues is used to populate
which OptionManager instance.

Follow up required for:
* VC to determine what the contents of adv opts should be.
* whether HostConfigInfo.Options is adv opts, or combined set